requirements.txt
tensorflow_macos==2.6.0
numpy==1.22.2
opencv_python==4.5.4.58
streamlit==1.5.0
Pillow==9.0.1
tensorflow==2.8.0
